InterCode/PlanAndSolve.py
```python
import numpy as np
from InterCode.intercode import InterCode

class PlanAndSolve:
    def __init__(self, intercode: InterCode):
        self.intercode = intercode

    def plan(self, task):
        # Plan the steps to solve the task
        # This is a placeholder and should be replaced with actual planning logic
        steps = ["step1", "step2", "step3"]
        return steps

    def solve(self, steps):
        # Execute the steps to solve the task
        # This is a placeholder and should be replaced with actual solving logic
        for step in steps:
            self.intercode.execute(step)

    def run(self, task):
        steps = self.plan(task)
        self.solve(steps)

if __name__ == "__main__":
    intercode = InterCode()
    plan_and_solve = PlanAndSolve(intercode)
    task = "Task to be solved"
    plan_and_solve.run(task)
```